Move compilation of TrueType hinting to outline compiler
